Add files via upload
Overlpar2.py contains the Window-wise overlap required in US simulations. The usage of the script goes as follows:
Firstly, place the script and all the dat files associated with each window of the CV.
Type in the command ls > input to store the names of all the dat files inside the file named "input".
Run the script by python3 overlap2.py input.