Usage of the Script
This is the code to run UMAP embedding on an MD trajectory. The protocol followed here is as follows:
Initially, hyperparameter tuning associated with UMAP embedding, such as min_distance and n_neighbours is done, using DBSCAN clustering and Silhouette coefficient as the metric.
Based on the SC value, best UMAP model is generated. 
The script needs two inputs, an nc file and a prmtop file. It will generate two csv files, one of them called "umap_silhouette_scores.csv" contains the SC value for each combination of min_distance and n_neighbours.
And secondly, umap_data_with_frame_numbers.csv" contains the value of UMAP-1 and UMAP-2 for each frame of the trajectory.